+++ include/H2DE/H2DE_renderer.h
@@ -12,6 +12,7 @@ private:
 
     std::unordered_map<std::string, SDL_Texture*> textures;
     std::unordered_map<std::string, Mix_Chunk*> sounds;
+    std::unordered_map<std::string, H2DE_Font> fonts;
 
     void destroyTextures();
     void destroySounds();
@@ -40,7 +41,10 @@ public:
     void render();
 
     friend void H2DE_LoadAssets(H2DE_Engine* engine, const std::filesystem::path& directory);
+    friend void H2DE_InitFont(H2DE_Engine* engine, const std::string& name, const H2DE_Font& font);
     H2DE_LevelPos absToLvl(const H2DE_AbsPos& pos, bool absolute) const;
+
+    const std::unordered_map<std::string, H2DE_Font>& getFonts() const;
 };
 
 #endif